// projects/ui/src/lib/components/po-skeleton/enums/po-skeleton-animation.enum.ts
/**
 * @usedBy PoSkeletonComponent, PoSkeletonContainerDirective
 *
 * @description
 *
 * Define os tipos de animação do componente `po-skeleton` e da diretiva `po-skeleton-container`.
 */
export enum PoSkeletonAnimation {
  /** Sem animação */
  none = 'none',

  /** Animação de pulsação */
  pulse = 'pulse',

  /** Animação de brilho deslizante */
  shimmer = 'shimmer'
}