// projects/ui/src/lib/components/po-skeleton/enums/po-skeleton-size.enum.ts
/**
 * @usedBy PoSkeletonComponent
 *
 * @description
 *
 * Define os tamanhos disponíveis para o componente `po-skeleton`.
 */
export enum PoSkeletonSize {
  /** Tamanho extra pequeno */
  xs = 'xs',

  /** Tamanho pequeno */
  sm = 'sm',

  /** Tamanho médio */
  md = 'md',

  /** Tamanho grande */
  lg = 'lg',

  /** Tamanho extra grande */
  xl = 'xl',

  /** Tamanho extra extra grande */
  xxl = '2xl'
}