// projects/ui/src/lib/components/po-skeleton/enums/po-skeleton-type.enum.ts
/**
 * @usedBy PoSkeletonComponent
 *
 * @description
 *
 * Define os tipos visuais disponíveis para o componente `po-skeleton`.
 */
export enum PoSkeletonType {
  /** Tipo padrão com cor neutra clara */
  normal = 'normal',

  /** Tipo primário com cor neutra média */
  primary = 'primary',

  /** Tipo de conteúdo com fundo branco */
  content = 'content'
}